RK2/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt
gunicorn==23.0.0
requests
pillow
python-telegram-bot==20.3